Added Home to navigation drawer. Renamed Expense to ExpenseAdapterItem. Changed ExpenseListFragment to read spending data from the database. Changed amount for Receipt to a Double. Changes to SubmitReceiptActivity: moved all findViewById calls to onCreate, changed Receipt amount to Double, changed Receipt date format, and changed Receipt user to displayName. Implemented master/detail flow in ViewSpendingActivity.

+++ mvlab7/app/src/main/java/com/example/myapplication/Receipt.java
@@ -1,7 +1,7 @@
 package com.example.myapplication;
 
 public class Receipt {
-    public Integer amount;
+    public Double amount;
     public String category;
     public Long date;
     public String description;
